refactor(front): API URL解決をresolveApiUrlに一本化する
context.cloudflareの手書きキャストとフォールバックが10箇所に散在していたのを
lib/api.tsのresolveApiUrl()に集約。movies.$idのクライアント側で本番ホスト
https://shine-api.yuta25.workers.dev を直書きしていた箇所も、loaderから
渡したapiUrlを使うように修正。
