+++ apps/front/app/routes/og-movie.tsx
@@ -7,6 +7,7 @@ import {
 import {fetchPosterAsDataUri, loadGoogleFont} from '@/lib/og/assets';
 import {OG_HEIGHT, OG_WIDTH, buildMovieCardHtml} from '@/lib/og/template';
 import {upgradePosterForSharing} from '@/lib/meta';
+import {resolveApiUrl} from '@/lib/api';
 
 type MovieDetail = {
   title?: string;
@@ -33,9 +34,7 @@ export async function loader({context, request}: Route.LoaderArgs) {
     return new Response('Not Found', {status: 404});
   }
 
-  const apiUrl =
-    (context.cloudflare as {env: {PUBLIC_API_URL?: string}}).env
-      .PUBLIC_API_URL ?? 'http://localhost:8787';
+  const apiUrl = resolveApiUrl(context);
 
   const response = await fetch(`${apiUrl}/movies/${id}`, {
     signal: request.signal,
